refactor out of main
