+++ src/js/electron/main.ts
@@ -22,10 +22,10 @@ import {windowsPre25Exists} from "./windows-pre-25"
 import {meta} from "src/app/ipc/meta"
 import secureWebContents from "./secure-web-contents"
 import env from "src/app/core/env"
-import path, {join} from "path"
+import {join} from "path"
 import requireAll from "./require-all"
 import isDev from "./isDev"
-import fs from "fs-extra"
+import migrateBrimToZui from "./migrateBrimToZui"
 require("@electron/remote/main").initialize()
 
 const pkg = meta.packageJSON()
@@ -42,54 +42,6 @@ export const mainDefaults = () => ({
 })
 
 export type MainArgs = ReturnType<typeof mainDefaults>
-
-const migrateBrimToZui = () => {
-  const zuiPath = app.getPath("userData")
-  const zuiAppStatePath = path.join(zuiPath, "appState.json")
-  const zuiAppLakePath = path.join(zuiPath, "lake")
-  const zuiAppDataPath = path.join(zuiPath, "data")
-
-  const brimPath = zuiPath.replace("Zui", "Brim")
-  if (zuiPath === brimPath) return
-
-  const brimAppStatePath = path.join(brimPath, "appState.json")
-  const brimAppLakePath = path.join(brimPath, "lake")
-  const brimAppDataPath = path.join(brimPath, "data")
-
-  try {
-    fs.statSync(brimAppStatePath)
-    fs.statSync(brimAppLakePath)
-    fs.statSync(brimAppDataPath)
-    log.info("brim data found")
-  } catch {
-    log.info("no brim data to migrate")
-    return
-  }
-
-  try {
-    fs.statSync(zuiAppStatePath)
-    fs.statSync(zuiAppLakePath)
-    fs.statSync(zuiAppDataPath)
-    log.info("zui data already exists, aborting migration")
-    return
-  } catch {
-    log.info("no existing zui data, proceeding to migrate")
-  }
-
-  try {
-    log.info(`migrating '${brimAppStatePath}' => '${zuiAppStatePath}'`)
-    fs.copySync(brimAppStatePath, zuiAppStatePath)
-    log.info(`migrating '${brimAppLakePath} => ${zuiAppLakePath}'`)
-    fs.copySync(brimAppLakePath, zuiAppLakePath)
-    log.info(`migrating '${brimAppDataPath} => ${zuiAppDataPath}'`)
-    fs.copySync(brimAppDataPath, zuiAppDataPath)
-  } catch (err) {
-    log.error("migration failed: ", err)
-    return
-  }
-
-  log.info("migration completed")
-}
 
 export async function main(args: Partial<MainArgs> = {}) {
   const opts = {...mainDefaults(), ...args}
